Changed int in packet_t to int16_t. Added lift, pitch, roll and yaw values to drone struct.

+++ in4073/command_types.h
@@ -24,5 +24,5 @@ struct packet_t
 {
 	// XXX: Add check bytes
 	char command;
-	int value;
+	int16_t value;
 };
