+++ in4073/pc_terminal/command_types.h
@@ -21,6 +21,7 @@
 
 struct packet_t
 {
+	// XXX: Add check bytes
 	char command;
-	int value;
+	int16_t value;
 };
